requirement.txt
Flask==3.0.3
beautifulsoup4==4.12.3
requests==2.32.3
google-generativeai==0.7.0
selenium==4.22.0
webdriver-manager==4.2.1
lxml==5.2.2